streamlit/requirements.txt
streamlit
requests
supabase
python-dotenv